Change/datatype

+++ pojo/src/main/java/com/stanzaliving/po/model/PoItemDetailsDto.java
@@ -47,7 +47,7 @@ public class PoItemDetailsDto {
 
     private BigDecimal shortPoApprovedQuantity;
 
-    private Float grnItemQuantity;
+    private Integer grnItemQuantity;
 
     private String hsnCode;
